CSV imported users can't authenticate
Fix #2082

When importing new users via CSV and electing to auto-generate a password for them and email is successfully sent but the links deliver application error 1901 and does't work. I've confirmed this on sandyjay instance.

Plugin was missing `token_set` method before sending the email with the generated hash, so it was never saved and there for invalid.